# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoRa KT API",
    "version": "1.0.0",
    "description": "REST API for team management, checkpoints, RFID access control, LoRa telemetry, and map data."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Relative to application root"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "webhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Optional server-level secret for the ingest endpoint. When LORA_WEBHOOK_SECRET is set (and not the default placeholder), every POST /api/ingest must include this header."
      }
    },
    "schemas": {
...
      "GroupCheckpoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "position"
        ]
      },
...
      "Checkpoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "easting": {
            "type": [
              "number",
              "null"
            ]
          },
          "northing": {
            "type": [
              "number",
              "null"
            ]
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupCheckpoint"
            }
          },
          "lora_device": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "id": {
                "type": "integer"
              },
              "dev_num": {
                "type": "integer"
              },
              "name": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/checkpoints": {
      "get": {
        "tags": [
          "checkpoints"
        ],
        "summary": "List checkpoints",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Checkpoints",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checkpoints": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Checkpoint"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "checkpoints"
        ],
        "summary": "Create checkpoint",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkpoint"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created checkpoint"
          }
        }
      }
    },
    "/api/checkpoints/{checkpoint_id}": {
      "parameters": [
        {
          "name": "checkpoint_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "checkpoints"
        ],
        "summary": "Checkpoint detail",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Checkpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkpoint"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "checkpoints"
        ],
        "summary": "Update checkpoint",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkpoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated checkpoint"
          }
        }
      },
      "delete": {
        "tags": [
          "checkpoints"
        ],
        "summary": "Delete checkpoint",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "put": {
        "tags": [
          "checkpoints"
        ],
        "summary": "Replace checkpoint",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Checkpoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced checkpoint"
          }
        }
      }
    },
    "/api/checkpoints/import": {
      "post": {
        "tags": [
          "checkpoints"
        ],
        "summary": "Bulk import checkpoints",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Checkpoint"
                    }
                  }
                },
                "required": [
                  "items"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import summary"
          }
        }
      }
    },
...
  }
}